fix: resolve remaining Kotlin compile errors across 6 files
- GameApiManager: getGameMode() needs package arg on API 31
- NetworkMonitor: suppress deprecated WifiInfo.connectionInfo + null safety
- NetworkAnalytics: explicit toFloat() for compound Float/Int assignment
- RoamingGuard: use literal string for ACTION_NETWORK_COUNTRY_CHANGED (API 31+)
- PlayIntegrityChecker: replace .await() with suspendCancellableCoroutine
- ServerHealthChecker: wrap async/awaitAll in coroutineScope {}

+++ gamevpn/app/src/main/kotlin/com/yourname/gamemodevpn/NetworkAnalytics.kt
@@ -100,7 +100,7 @@ class HeatmapView(ctx: Context) : View(ctx) {
             counts[day][hour]++
         }
         for (d in 0..6) for (h in 0..23)
-            if (counts[d][h] > 0) newData[d][h] /= counts[d][h]
+            if (counts[d][h] > 0) newData[d][h] /= counts[d][h].toFloat()
         data = newData
         invalidate()
     }
